Add C# tests
