+++ src/Razor/test/Microsoft.AspNetCore.Razor.LanguageServer.Test/Completion/DefaultTagHelperCompletionServiceTest.cs
@@ -3,115 +3,17 @@
 
 using System;
 using Microsoft.AspNetCore.Razor.Language;
-using Microsoft.AspNetCore.Razor.Language.Components;
 using Microsoft.AspNetCore.Razor.Language.Legacy;
 using Microsoft.AspNetCore.Razor.Language.Syntax;
-using Microsoft.AspNetCore.Razor.Test.Common;
 using Microsoft.VisualStudio.Editor.Razor;
 using OmniSharp.Extensions.LanguageServer.Protocol.Models;
 using Xunit;
-using DefaultRazorTagHelperCompletionService = Microsoft.VisualStudio.Editor.Razor.DefaultTagHelperCompletionService;
-using RazorTagHelperCompletionService = Microsoft.VisualStudio.Editor.Razor.TagHelperCompletionService;
 
 namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
 {
-    public class DefaultTagHelperCompletionServiceTest : LanguageServerTestBase
+
+    public class DefaultTagHelperCompletionServiceTest : DefaultTagHelperServiceTestBase
     {
-        public DefaultTagHelperCompletionServiceTest()
-        {
-            var builder1 = TagHelperDescriptorBuilder.Create("Test1TagHelper", "TestAssembly");
-            builder1.TagMatchingRule(rule => rule.TagName = "test1");
-            builder1.SetTypeName("Test1TagHelper");
-            builder1.BindAttribute(attribute =>
-            {
-                attribute.Name = "bool-val";
-                attribute.SetPropertyName("BoolVal");
-                attribute.TypeName = typeof(bool).FullName;
-            });
-            builder1.BindAttribute(attribute =>
-            {
-                attribute.Name = "int-val";
-                attribute.SetPropertyName("IntVal");
-                attribute.TypeName = typeof(int).FullName;
-            });
-
-            var builder2 = TagHelperDescriptorBuilder.Create("Test2TagHelper", "TestAssembly");
-            builder2.TagMatchingRule(rule => rule.TagName = "test2");
-            builder2.SetTypeName("Test2TagHelper");
-            builder2.BindAttribute(attribute =>
-            {
-                attribute.Name = "bool-val";
-                attribute.SetPropertyName("BoolVal");
-                attribute.TypeName = typeof(bool).FullName;
-            });
-            builder2.BindAttribute(attribute =>
-            {
-                attribute.Name = "int-val";
-                attribute.SetPropertyName("IntVal");
-                attribute.TypeName = typeof(int).FullName;
-            });
-
-            var directiveAttribute1 = TagHelperDescriptorBuilder.Create(ComponentMetadata.Component.TagHelperKind, "TestDirectiveAttribute", "TestAssembly");
-            directiveAttribute1.TagMatchingRule(rule =>
-            {
-                rule.TagName = "*";
-            });
-            directiveAttribute1.BindAttribute(attribute =>
-            {
-                attribute.Metadata[ComponentMetadata.Common.DirectiveAttribute] = bool.TrueString;
-                attribute.Name = "@test";
-                attribute.SetPropertyName("Test");
-                attribute.TypeName = typeof(string).FullName;
-
-                attribute.BindAttributeParameter(parameter =>
-                {
-                    parameter.Name = "something";
-                    parameter.TypeName = typeof(string).FullName;
-
-                    parameter.SetPropertyName("Something");
-                });
-            });
-            directiveAttribute1.Metadata[ComponentMetadata.Component.NameMatchKey] = ComponentMetadata.Component.FullyQualifiedNameMatch;
-            directiveAttribute1.SetTypeName("TestDirectiveAttribute");
-
-            var directiveAttribute2 = TagHelperDescriptorBuilder.Create(ComponentMetadata.Component.TagHelperKind, "MinimizedDirectiveAttribute", "TestAssembly");
-            directiveAttribute2.TagMatchingRule(rule =>
-            {
-                rule.TagName = "*";
-            });
-            directiveAttribute2.BindAttribute(attribute =>
-            {
-                attribute.Metadata[ComponentMetadata.Common.DirectiveAttribute] = bool.TrueString;
-                attribute.Name = "@minimized";
-                attribute.SetPropertyName("Minimized");
-                attribute.TypeName = typeof(bool).FullName;
-
-                attribute.BindAttributeParameter(parameter =>
-                {
-                    parameter.Name = "something";
-                    parameter.TypeName = typeof(string).FullName;
-
-                    parameter.SetPropertyName("Something");
-                });
-            });
-            directiveAttribute2.Metadata[ComponentMetadata.Component.NameMatchKey] = ComponentMetadata.Component.FullyQualifiedNameMatch;
-            directiveAttribute2.SetTypeName("TestDirectiveAttribute");
-
-            DefaultTagHelpers = new[] { builder1.Build(), builder2.Build(), directiveAttribute1.Build() };
-            var tagHelperFactsService = new DefaultTagHelperFactsService();
-            RazorTagHelperCompletionService = new DefaultRazorTagHelperCompletionService(tagHelperFactsService);
-            HtmlFactsService = new DefaultHtmlFactsService();
-            TagHelperFactsService = new DefaultTagHelperFactsService();
-        }
-
-        private TagHelperDescriptor[] DefaultTagHelpers { get; }
-
-        private RazorTagHelperCompletionService RazorTagHelperCompletionService { get; }
-
-        private HtmlFactsService HtmlFactsService { get; }
-
-        private TagHelperFactsService TagHelperFactsService { get; }
-
         [Fact]
         public void GetNearestAncestorTagInfo_MarkupElement()
         {
@@ -148,231 +50,6 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
             // Assert
             Assert.Equal("test1", ancestorName);
             Assert.True(ancestorIsTagHelper);
-        }
-
-        [Fact]
-        public void StringifyAttributes_DirectiveAttribute()
-        {
-            // Arrange
-            var codeDocument = CreateComponentDocument($"<TestElement @test='abc' />", DefaultTagHelpers);
-            var syntaxTree = codeDocument.GetSyntaxTree();
-            var sourceSpan = new SourceSpan(3, 0);
-            var sourceChangeLocation = new SourceChange(sourceSpan, string.Empty);
-            var startTag = (MarkupTagHelperStartTagSyntax)syntaxTree.Root.LocateOwner(sourceChangeLocation).Parent;
-
-            // Act
-            var attributes = TagHelperFactsService.StringifyAttributes(startTag.Attributes);
-
-            // Assert
-            Assert.Collection(
-                attributes,
-                attribute =>
-                {
-                    Assert.Equal("@test", attribute.Key);
-                    Assert.Equal("abc", attribute.Value);
-                });
-        }
-
-        [Fact]
-        public void StringifyAttributes_DirectiveAttributeWithParameter()
-        {
-            // Arrange
-            var codeDocument = CreateComponentDocument($"<TestElement @test:something='abc' />", DefaultTagHelpers);
-            var syntaxTree = codeDocument.GetSyntaxTree();
-            var sourceSpan = new SourceSpan(3, 0);
-            var sourceChangeLocation = new SourceChange(sourceSpan, string.Empty);
-            var startTag = (MarkupTagHelperStartTagSyntax)syntaxTree.Root.LocateOwner(sourceChangeLocation).Parent;
-
-            // Act
-            var attributes = TagHelperFactsService.StringifyAttributes(startTag.Attributes);
-
-            // Assert
-            Assert.Collection(
-                attributes,
-                attribute =>
-                {
-                    Assert.Equal("@test:something", attribute.Key);
-                    Assert.Equal("abc", attribute.Value);
-                });
-        }
-
-        [Fact]
-        public void StringifyAttributes_MinimizedDirectiveAttribute()
-        {
-            // Arrange
-            var codeDocument = CreateComponentDocument($"<TestElement @minimized />", DefaultTagHelpers);
-            var syntaxTree = codeDocument.GetSyntaxTree();
-            var sourceSpan = new SourceSpan(3, 0);
-            var sourceChangeLocation = new SourceChange(sourceSpan, string.Empty);
-            var startTag = (MarkupTagHelperStartTagSyntax)syntaxTree.Root.LocateOwner(sourceChangeLocation).Parent;
-
-            // Act
-            var attributes = TagHelperFactsService.StringifyAttributes(startTag.Attributes);
-
-            // Assert
-            Assert.Collection(
-                attributes,
-                attribute =>
-                {
-                    Assert.Equal("@minimized", attribute.Key);
-                    Assert.Equal(string.Empty, attribute.Value);
-                });
-        }
-
-        [Fact]
-        public void StringifyAttributes_MinimizedDirectiveAttributeWithParameter()
-        {
-            // Arrange
-            var codeDocument = CreateComponentDocument($"<TestElement @minimized:something />", DefaultTagHelpers);
-            var syntaxTree = codeDocument.GetSyntaxTree();
-            var sourceSpan = new SourceSpan(3, 0);
-            var sourceChangeLocation = new SourceChange(sourceSpan, string.Empty);
-            var startTag = (MarkupTagHelperStartTagSyntax)syntaxTree.Root.LocateOwner(sourceChangeLocation).Parent;
-
-            // Act
-            var attributes = TagHelperFactsService.StringifyAttributes(startTag.Attributes);
-
-            // Assert
-            Assert.Collection(
-                attributes,
-                attribute =>
-                {
-                    Assert.Equal("@minimized:something", attribute.Key);
-                    Assert.Equal(string.Empty, attribute.Value);
-                });
-        }
-
-        [Fact]
-        public void StringifyAttributes_TagHelperAttribute()
-        {
-            // Arrange
-            var tagHelper = TagHelperDescriptorBuilder.Create("WithBoundAttribute", "TestAssembly");
-            tagHelper.TagMatchingRule(rule => rule.TagName = "test");
-            tagHelper.BindAttribute(attribute =>
-            {
-                attribute.Name = "bound";
-                attribute.SetPropertyName("Bound");
-                attribute.TypeName = typeof(bool).FullName;
-            });
-            tagHelper.SetTypeName("WithBoundAttribute");
-            var codeDocument = CreateCodeDocument($"@addTagHelper *, TestAssembly{Environment.NewLine}<test bound='true' />", tagHelper.Build());
-            var syntaxTree = codeDocument.GetSyntaxTree();
-            var sourceSpan = new SourceSpan(30 + Environment.NewLine.Length, 0);
-            var sourceChangeLocation = new SourceChange(sourceSpan, string.Empty);
-            var startTag = (MarkupTagHelperStartTagSyntax)syntaxTree.Root.LocateOwner(sourceChangeLocation).Parent;
-
-            // Act
-            var attributes = TagHelperFactsService.StringifyAttributes(startTag.Attributes);
-
-            // Assert
-            Assert.Collection(
-                attributes,
-                attribute =>
-                {
-                    Assert.Equal("bound", attribute.Key);
-                    Assert.Equal("true", attribute.Value);
-                });
-        }
-
-        [Fact]
-        public void StringifyAttributes_MinimizedTagHelperAttribute()
-        {
-            // Arrange
-            var tagHelper = TagHelperDescriptorBuilder.Create("WithBoundAttribute", "TestAssembly");
-            tagHelper.TagMatchingRule(rule => rule.TagName = "test");
-            tagHelper.BindAttribute(attribute =>
-            {
-                attribute.Name = "bound";
-                attribute.SetPropertyName("Bound");
-                attribute.TypeName = typeof(bool).FullName;
-            });
-            tagHelper.SetTypeName("WithBoundAttribute");
-            var codeDocument = CreateCodeDocument($"@addTagHelper *, TestAssembly{Environment.NewLine}<test bound />", tagHelper.Build());
-            var syntaxTree = codeDocument.GetSyntaxTree();
-            var sourceSpan = new SourceSpan(30 + Environment.NewLine.Length, 0);
-            var sourceChangeLocation = new SourceChange(sourceSpan, string.Empty);
-            var startTag = (MarkupTagHelperStartTagSyntax)syntaxTree.Root.LocateOwner(sourceChangeLocation).Parent;
-
-            // Act
-            var attributes = TagHelperFactsService.StringifyAttributes(startTag.Attributes);
-
-            // Assert
-            Assert.Collection(
-                attributes,
-                attribute =>
-                {
-                    Assert.Equal("bound", attribute.Key);
-                    Assert.Equal(string.Empty, attribute.Value);
-                });
-        }
-
-        [Fact]
-        public void StringifyAttributes_UnboundAttribute()
-        {
-            // Arrange
-            var codeDocument = CreateCodeDocument($"@addTagHelper *, TestAssembly{Environment.NewLine}<input unbound='hello world' />", DefaultTagHelpers);
-            var syntaxTree = codeDocument.GetSyntaxTree();
-            var sourceSpan = new SourceSpan(30 + Environment.NewLine.Length, 0);
-            var sourceChangeLocation = new SourceChange(sourceSpan, string.Empty);
-            var startTag = (MarkupStartTagSyntax)syntaxTree.Root.LocateOwner(sourceChangeLocation).Parent;
-
-            // Act
-            var attributes = TagHelperFactsService.StringifyAttributes(startTag.Attributes);
-
-            // Assert
-            Assert.Collection(
-                attributes,
-                attribute =>
-                {
-                    Assert.Equal("unbound", attribute.Key);
-                    Assert.Equal("hello world", attribute.Value);
-                });
-        }
-
-        [Fact]
-        public void StringifyAttributes_UnboundMinimizedAttribute()
-        {
-            // Arrange
-            var codeDocument = CreateCodeDocument($"@addTagHelper *, TestAssembly{Environment.NewLine}<input unbound />", DefaultTagHelpers);
-            var syntaxTree = codeDocument.GetSyntaxTree();
-            var sourceSpan = new SourceSpan(30 + Environment.NewLine.Length, 0);
-            var sourceChangeLocation = new SourceChange(sourceSpan, string.Empty);
-            var startTag = (MarkupStartTagSyntax)syntaxTree.Root.LocateOwner(sourceChangeLocation).Parent;
-
-            // Act
-            var attributes = TagHelperFactsService.StringifyAttributes(startTag.Attributes);
-
-            // Assert
-            Assert.Collection(
-                attributes,
-                attribute =>
-                {
-                    Assert.Equal("unbound", attribute.Key);
-                    Assert.Equal(string.Empty, attribute.Value);
-                });
-        }
-
-        [Fact]
-        public void StringifyAttributes_IgnoresMiscContent()
-        {
-            // Arrange
-            var codeDocument = CreateCodeDocument($"@addTagHelper *, TestAssembly{Environment.NewLine}<input unbound @DateTime.Now />", DefaultTagHelpers);
-            var syntaxTree = codeDocument.GetSyntaxTree();
-            var sourceSpan = new SourceSpan(30 + Environment.NewLine.Length, 0);
-            var sourceChangeLocation = new SourceChange(sourceSpan, string.Empty);
-            var startTag = (MarkupStartTagSyntax)syntaxTree.Root.LocateOwner(sourceChangeLocation).Parent;
-
-            // Act
-            var attributes = TagHelperFactsService.StringifyAttributes(startTag.Attributes);
-
-            // Assert
-            Assert.Collection(
-                attributes,
-                attribute =>
-                {
-                    Assert.Equal("unbound", attribute.Key);
-                    Assert.Equal(string.Empty, attribute.Value);
-                });
         }
 
         [Fact]
@@ -675,24 +352,6 @@ namespace Microsoft.AspNetCore.Razor.LanguageServer.Completion
                 completions,
                 completion => Assert.Equal("bool-val", completion.FilterText),
                 completion => Assert.Equal("int-val", completion.FilterText));
-        }
-
-        private static RazorCodeDocument CreateCodeDocument(string text, params TagHelperDescriptor[] tagHelpers)
-        {
-            tagHelpers = tagHelpers ?? Array.Empty<TagHelperDescriptor>();
-            var sourceDocument = TestRazorSourceDocument.Create(text);
-            var projectEngine = RazorProjectEngine.Create(builder => { });
-            var codeDocument = projectEngine.ProcessDesignTime(sourceDocument, "mvc", Array.Empty<RazorSourceDocument>(), tagHelpers);
-            return codeDocument;
-        }
-
-        private static RazorCodeDocument CreateComponentDocument(string text, params TagHelperDescriptor[] tagHelpers)
-        {
-            tagHelpers = tagHelpers ?? Array.Empty<TagHelperDescriptor>();
-            var sourceDocument = TestRazorSourceDocument.Create(text);
-            var projectEngine = RazorProjectEngine.Create(builder => { });
-            var codeDocument = projectEngine.ProcessDesignTime(sourceDocument, FileKinds.Component, Array.Empty<RazorSourceDocument>(), tagHelpers);
-            return codeDocument;
         }
     }
 }
